C60 piece 3: hash iteration via Ordered on=3
Hash iteration (`for e in h { ... }`) now reuses the existing Ordered
(on=3) iteration path — same bytecode, same element type
(`reference<T>`), same loop attributes as sorted/index.

Three small edits wire it up:

- parser/collections.rs parse_for: when `in_type` is `Type::Hash`,
  substitute the iterated expression with a scratch variable holding
  the result of `n_hash_sorted(h, tp)` — a u32-stride rec-nr vector
  built in the hash's own store (edit A landed earlier in 2e20ba2).
  Scratch variable typed as `Type::Reference(content, dep)` so downstream
  size/free machinery resolves the struct definition correctly.
- parser/collections.rs iterator(): extend the elem_type match to
  include `Type::Hash` so the loop-body iter_next block is annotated
  with `Reference<T>` instead of falling through to `Type::Null`
  (which caused field access on the loop variable to return null).
- parser/fields.rs fill_iter Hash arm: emit `on=3, arg=4`. The
  scratch's layout matches Ordered's exactly — Ordered iteration
  yields `DbRef{store=hash_store, rec, pos=8}`, which resolves to a
  valid hash record body because the scratch lives in the hash's
  store (edit A).

Five of six C60 integration tests now pass
(c60_hash_iter_single_field_asc, _loop_attributes, _multi_field_lex,
_filter_clause, _empty); the sixth (`_remove_rejected`) stays
ignored pending the Step 9 `#remove` diagnostic.

+++ src/state/io.rs
@@ -669,14 +669,15 @@ impl State {
                     // C60 piece 3 edit E: unbounded iteration (`for e
                     // in h { … }` with no range).  ordered_find with an
                     // empty key returns (0, true) which collapses
-                    // start=0 and finish=0, so the step protocol never
-                    // fires even once.  Set start to the u32::MAX
-                    // sentinel so vector_next's first call resets pos
-                    // to 8 (first element); finish is unused for on=3
-                    // (termination is via vector_next returning MAX
-                    // when past the last element).  This mirrors the
-                    // explicit is_empty() guard Sorted (on=2) has.
-                    start = u32::MAX;
+                    // start=0/finish=0, so the step protocol never
+                    // fires even once.  Set start to the "not started"
+                    // sentinel that `vector_next` recognises at
+                    // src/vector.rs:455 — it checks `*pos == i32::MAX`
+                    // (the i32 positive max, 0x7FFF_FFFF), *not*
+                    // u32::MAX.  Passing u32::MAX here casts to i32 as
+                    // -1 and falls into the "advance" branch, reading
+                    // garbage at pos=-1+size.
+                    start = i32::MAX as u32;
                     finish = 0;
                 } else if reverse {
                     start = vector::ordered_find(&data, true, all, &keys, &from).0 + u32::from(!ex);
